% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgbob.R
\name{svgbob}
\alias{svgbob}
\title{SvgBob}
\usage{
svgbob(text, svgfile = tempfile())
}
\arguments{
\item{text}{input ascii text}

\item{svgfile}{output file to write svg}
}
\description{
Converts ascii art into beautiful svg images.
}
\examples{
text <- readLines(system.file('examples/circuits.bob', package = 'svgbob'))
svgbob(text[1:41], 'circuits.svg')
utils::browseURL('circuits.svg')

# Convert to png
rsvg::rsvg_png('circuits.svg', 'circuits.png', width = 2000)
utils::browseURL('circuits.png')
}
\references{
Live demo: \url{http://ivanceras.github.io/svgbob-editor/}
}
